Update RunModel.R
Hi Dr. Willoughby,

I only slightly changed your template when playing around with the model to add a few column names and an output table. I also had to change "nrow(intransit)" to "length(intransit)" because it was a vector. 

Ultimately, I think it is best that the starting point be assuming some population of mosquitos are on board the ship such as defined in this model as "wildtocargo." 

I am also not very familiar with GitHub so let me know if there are other questions or changes I can answer.

Thank you!
Jordan